feat(api): organizando camada de requests
