+++ api/app/Http/Requests/course/UpdateCourseRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\course;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class UpdateKeywordRequest extends FormRequest
+class UpdateCourseRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
