+++ api/app/Http/Requests/description/UpdateDescriptionRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\description;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class UpdateCourseRequest extends FormRequest
+class UpdateDescriptionRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
