+++ api/app/Http/Requests/example/StoreExampleRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\example;
 
 use Illuminate\Foundation\Http\FormRequest;
