+++ api/app/Http/Requests/example/UpdateExampleRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\example;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class UpdateDescriptionRequest extends FormRequest
+class UpdateExampleRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
