+++ api/app/Http/Requests/favorite/StoreFavoriteRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\favorite;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class StoreSignRequest extends FormRequest
+class StoreFavoriteRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
