+++ api/app/Http/Requests/favorite/UpdateFavoriteRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\favorite;
 
 use Illuminate\Foundation\Http\FormRequest;
