+++ api/app/Http/Requests/keyword/UpdateKeywordRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\keyword;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class StoreCourseRequest extends FormRequest
+class UpdateKeywordRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
